Allow saving same credentials for all ports on host
If the auth requiring service does not run on a standart port (i. e. 80
for http), the user is asked if the credentials should stored for all
ports. It is still possible to specify credentials for a single port
(or ignore a single port). When providing credentials, first the
settings for the host with a specific port are checked, then the
settings for the generic host are taken (if any exist).

Resolves #3.